requirements.txt
numpy==1.24.1
pandas==1.5.2
Pillow==9.4.0
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0